aegis_core/requirements.txt
# AegisOps Agent Core dependencies
# Python 3.10+

fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
google-generativeai>=0.5,<1.0
docker>=7.0,<8.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
